Move packet receiving and buffering functions to separate module
